Interpret Gerrit message timestamps in UTC.
Gerrit returns message timestamps in UTC. However, we currently parse
the timestamp as a local time, resulting in messages being marked
stale too early or late for users in non-UTC timezones.

Update message timestamp parsing to use UTC. Additionally, change the
name "getDate" to "getTime" to make it clearer the returned value is
a timestamp.