Improve stale CL classification
If a CL is uploaded with `git cl upload --reviewer xxx@example.com`,
then it has a reviewer assigned but all messages are auto-generated.
Previously, such CLs where considered as stale (since there were no
messages that could be used to determine the date of last updated).

Instead, if all messages are auto-generated, then use the timestamp
of the last such messages to determine whether the CL is stale or
not.

This should fix issue #3.