Show information for reachable hosts even if some are unreachable
If some hosts are reachable, then display the information about
those gerrit instance in addition to a button to log into the
ones that are not (assuming that the user need to refresh those).

This is implemented by changing fetchCls to use Promise.allSettled
(which wait until all promise are resolved or rejected and then
return a list of promise results) instead of Promise.all (that is
rejected if any of the Promise is rejected).

The list of promise results is then converted into an object that
contains a gerrit.SearchResults() and a list of errors (possibly
empty). The function popup.displayPopup then use the data to
present as much information to the user as possible.

When some of the hosts are unreachable, then a button will be
presented to log in. The display is a bit ugly, but functional.

Fixes issue #7.