Differentiate between CLs without reviewers and WIPs
There should be a distinction between CLs that have been sent out for
review but that do not have assigned reviewers, and those that are
published to Gerrit with the "work in progress" label attached.

This change splits these distinct states into separate categories with
different labels, but they still use the same grey icon as before (now
renamed to "not_ready").