Fetch all required information from Gerrit for background refresh.
Currently, background refreshes of data from Gerrit silently fail due to
the error:

    TypeError: Cannot read property 'undefined' of undefined
        at Changelist.getCurrentRevision (gerrit.js:268)
        at Changelist.getCategory (gerrit.js:191)
        at gerrit.js:411
        at Array.forEach (<anonymous>)
        at SearchResult.getCategoryMap (gerrit.js:410)
        at gerrit.js:452
        at Array.forEach (<anonymous>)
        at SearchResults.getCategoryMap (gerrit.js:451)
        at update (badge.js:94)
        at badge.js:46

This is triggered because |getCurrentRevision| requires CURRENT_REVISION
to be sent in the request to Gerrit, but the background badge refresh
requests a non-detailed fetch of information, which doesn't include
this.

This CL adds CURRENT_REVISION to the set of non-detailed information,
resolving the above error.

I've kept the current detailed/non-detailed split, because requesting
DETAILED_ACCOUNTS roughly triples the payload size in my (very brief)
experiments.