Fix "attention set only" option
The "options" object passed to the SearchResult constructor was a promise
in some cases and thus accessing the property always returned "undefined"
when checking whether the option was enabled.

Fix this by waiting for the browser.loadOptions() promise to be resolved.
Since fetching the list of instances also requires loading the options,
change gerrit.fetchAllowedInstances() and gerrit.fetchAllInstances() to
take the options as a parameter (to avoid waiting multiple time for the
promise to be resolved).

Fix the code reading the property from "options" to use the correct name
and change the code to compare the value to config.OPTION_ENABLED (as it
is a non-emptry string and not a boolean thus cannot be used in a boolean
expression).

Fixes #30.