Fix saving of experimental settings value
The code presenting the extension configuration had a bug where the
old value for the experimental settings was read to initialize the
html widget but not saved into the "Options" object. This caused it
to revert the value back to "unspecified" (the default) if the user
opened the page and saved the options without changing those values.

Fix this by recording the loaded values into "Options" to avoid the
settings reverting back to "unspecified" when the user make changes
to the options.

Fixes #33.