Request all permissions at the same time
The CL introducing the request for the "notifications" permissions
put the request in a separate call to chrome.permissions.request.

For user that had already configured gerrit instance and just went
to the option page to enable the notification, a single popup was
presented, but for those that changed the list of instances at the
same time, two pop-up were presented.

By merging the request into a single call, user will be presented
with a single popup showing all permissions requested. This make
the experience a bit better for new user (as "notifications" are
now enabled by default).