Cache list of reviews/errors from Gerrit hosts
Opening the popup can be slow if the user has many hosts or CLs
in review. Cache the data collected from the gerrit instances
in memory for two and a half minute.

This should improve the performance from opening the popup while
still giving fresh results (assuming that the user wants to open
a few CLs).

The cache is also populated by the background refresh (happens
every five minutes) so the cache should have a high hit rate.

Fixes #17.