Sort the CLs per last-updated timestamp
Previously the CLs were presented in an arbitrary order (partitioned
per gerrit instance, then in the order the server returned the result).

Instead change the order to sort them according to the "updated"
timestamp (which is set by the server). This should ensure that the
CLs at the top of each category is the most recent CL, indendently
of which gerrit instance is hosting that CL.

Fixes #27.