Add URL parameter informing that request comes from gerrit-monitor
The administrator of Gerrit instances may want to know how many requests
comes from gerrit-monitor. Add a 'usp' search parameter when opening a
new tab from the popup menu (the server will redirect to the URL without
the parameter, thus use the bare URL when checking for an existing tab).