Add "Open all" option in pop-up for sections with multiple CLs
Add a new "Open all" label that will open all the CLs in the
current section (as if the user clicked on all of them, but
without re-using existing tabs as I found it confusing).

Add a new "link" class that changes the cursor to the "pointer"
style making it easier to understand that "Open all" and the
individual CLs are clickable.

Fixes #28.